Move field reflection into a shared ReflectionUtils
The read/writeField helpers lived on the home test base, where nothing else
could reuse them. Move them to org.commcare.android.util.ReflectionUtils.

Two of the call sites didn't need reflection at all: the dialogs a launch check
raises are stashed only because fragments are paused during create, so the tests
now ask the activity to show what it stashed — as onResume does — and assert on
the real dialog fragment.

The instance-state round-trip keeps reflection. Those four flags are private,
have no accessors, and reach no view or intent, so there is nothing to observe
them through; the alternative is dropping save/restore coverage on the eve of
the refactor that moves them.

[AI]
